+++ app/src/main/java/woowacourse/payments/ui/common/ExtraKeys.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.common
 
 object ExtraKeys {
     const val CARD_KEY = "card_key"
