+++ app/src/main/java/woowacourse/payments/ui/common/IntentCompat.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.common
 
 import android.content.Intent
 import android.os.Build
