+++ app/src/main/java/woowacourse/payments/ui/common/composable/PaymentCard.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.common
+package woowacourse.payments.ui.common.composable
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
